fix: use proper ConnectButtons component
